+++ Bud.Core/src/main/cs/Cli/Command.cs
@@ -1,0 +1,7 @@
+using Bud.Commander;
+
+namespace Bud.Cli {
+  public abstract class Command {
+    public abstract string EvaluateToJson(IBuildCommander buildCommander);
+  }
+}