+++ Bud.Core/src/main/cs/Commander/BuildCommanderExtensions.cs
@@ -13,5 +13,9 @@ namespace Bud.Commander {
     public static T Evaluate<T>(this IBuildCommander buildCommander, string commandString) {
       return JsonConvert.DeserializeObject<T>(buildCommander.EvaluateToJson(commandString));
     }
+
+    public static T EvaluateMacro<T>(this IBuildCommander buildCommander, string macroName, params string[] parameters) {
+      return JsonConvert.DeserializeObject<T>(buildCommander.EvaluateMacroToJson(macroName, parameters));
+    }
   }
 }