+++ Bud.Core/src/main/cs/Commander/DefaultBuildCommander.cs
@@ -1,24 +1,21 @@
 using Bud.Build;
-using Bud.IO;
 using Bud.Logging;
 
 namespace Bud.Commander {
   public class DefaultBuildCommander : IBuildCommander {
-    private readonly Settings Settings;
-    private readonly Config Config;
+    private BuildContext BuildContext;
 
     public DefaultBuildCommander(string dirOfProjectToBeBuilt, bool isQuiet)
       : this(CreateDefaultSettings(dirOfProjectToBeBuilt), isQuiet) {}
 
     public DefaultBuildCommander(Settings settings, bool isQuiet) {
-      Settings = settings;
-      Config = new Config(settings.ConfigDefinitions, isQuiet ? Logger.NullLogger : Logger.CreateFromStandardOutputs());
+      var logger = isQuiet ? Logger.NullLogger : Logger.CreateFromStandardOutputs();
+      BuildContext = new BuildContext(settings, logger);
     }
 
-    public string EvaluateToJson(string command) {
-      var context = Context.FromConfig(Config, Settings.TaskDefinitions);
-      return CommandEvaluator.EvaluateToJsonSynchronously(context, command);
-    }
+    public string EvaluateToJson(string command) => CommandEvaluator.EvaluateToJsonSync(command, ref BuildContext);
+
+    public string EvaluateMacroToJson(string macroName, params string[] commandLineParameters) => CommandEvaluator.EvaluateMacroToJsonSync(macroName, commandLineParameters, ref BuildContext);
 
     public void Dispose() {}