+++ Bud.Core/src/main/cs/MacroFunction.cs
@@ -1,0 +1,3 @@
+namespace Bud {
+  public delegate MacroResult MacroFunction(BuildContext context, string[] commandLineArguments);
+}